+++ README.md
@@ -71,12 +71,12 @@ You can run scenarios against a given base as:
 
 #### New Scenarios
 
-To add a new scenario, add to `test/scenarios`, e.g.
+To add a new scenario, add to `scenario/`, e.g.
 
-**tests/scenarios/NewToken.ts**
+**scenario/NewToken.ts**
 
 ```ts
-import { scenario, World } from '../scen2';
+import { scenario, World } from '../plugins/scenario';
 import { CometContext } from './Context';
 import { expect } from 'chai';
