+++ plugins/scenario/worker/BootstrapWorker.js
