+++ plugins/scenario/worker/Config.ts
@@ -1,6 +1,6 @@
 
 // TODO: This doesn't belong here
 
-export const scenarioGlob = "test/scenarios/**.ts";
+export const scenarioGlob = "scenario/**.ts";
 export const workerCount = 3;
 export const defaultFormats = ["console"];
