+++ tasks/spider/task.ts
@@ -1,7 +1,7 @@
 import { task } from 'hardhat/config';
-import { pullConfigs } from './spider';
+import { pullConfigs } from '../../plugins/spider';
 
 task('spider', 'Use Spider method to pull in contract configs')
   .setAction(async (taskArgs, hre) => {
     await pullConfigs(hre);
-  })
+  });
